feat: typscript migration

+++ tests/utils/test-helpers.ts
@@ -32,7 +32,9 @@ export async function clearTestDatabase(): Promise<void> {
   
   for (const key in collections) {
     const collection = collections[key];
-    await collection.deleteMany({});
+    if (collection) {
+      await collection.deleteMany({});
+    }
   }
 }
